# api/agent/swagger.yaml
basePath: /v1
definitions:
  api.appSessionResponse:
    properties:
      app_id:
        type: string
      issued_at:
        type: string
      token:
        type: string
      workflows_run:
        items:
          type: string
        type: array
    type: object
  api.appSummary:
    properties:
      id:
        type: string
      version:
        type: string
      workflows_run:
        items:
          type: string
        type: array
    type: object
  api.connectorSecretRef:
    properties:
      key:
        type: string
      type:
        type: string
    type: object
  api.connectorSummary:
    properties:
      config:
        additionalProperties: {}
        type: object
      created_at:
        type: string
      id:
        type: string
      secret_refs:
        additionalProperties:
          $ref: '#/definitions/api.connectorSecretRef'
        type: object
      type:
        type: string
    type: object
  api.connectorTestResponse:
    properties:
      message:
        type: string
      ok:
        type: boolean
    type: object
  api.createConnectorRequest:
    properties:
      config:
        additionalProperties: {}
        type: object
      id:
        type: string
      secret_refs:
        additionalProperties:
          $ref: '#/definitions/api.connectorSecretRef'
        type: object
      type:
        type: string
    type: object
...
  api.pluginSummary:
    properties:
      actions:
        items:
          type: string
        type: array
      connectors:
        items:
          type: string
        type: array
      id:
        type: string
      version:
        type: string
    type: object
  api.runStep:
    properties:
      error:
        type: string
      finished_at:
        type: string
      id:
        type: string
      input:
        additionalProperties: {}
        type: object
      output:
        additionalProperties: {}
        type: object
      started_at:
        type: string
      status:
        type: string
    type: object
  api.runSummary:
    properties:
      created_at:
        type: string
      error:
        type: string
      finished_at:
        type: string
      id:
        type: string
      inputs:
        additionalProperties: {}
        type: object
      outputs:
        additionalProperties: {}
        type: object
      started_at:
        type: string
      status:
        type: string
      steps:
        items:
          $ref: '#/definitions/api.runStep'
        type: array
      workflow_id:
        type: string
      workflow_version:
        type: integer
    type: object
  api.runWorkflowRequest:
    properties:
      bindings:
        additionalProperties:
          type: string
        type: object
      inputs:
        additionalProperties: {}
        type: object
    type: object
  api.workflowBindingDetail:
    properties:
      connector_type:
        description: |-
          ConnectorType is empty when it could not be inferred for every step
          that references this binding name, or when different steps referring
          to the same name disagree on it.
        type: string
      name:
        type: string
    type: object
  api.workflowDetail:
    properties:
      bindings:
        description: |-
          Bindings is every distinct connector binding name referenced by Steps
          (via a "${{ bindings.<name> }}" expression), each paired with its
          inferred connector type when one could be — see bindingConnectorType.
        items:
          $ref: '#/definitions/api.workflowBindingDetail'
        type: array
      id:
        type: string
      inputs:
        items:
          $ref: '#/definitions/api.workflowInputDetail'
        type: array
      next_run_at:
        description: |-
          NextRunAt is this workflow's next scheduled firing, populated from
          the schedules table (internal/scheduler) rather than parsed from the
          definition — nil for a "manual" trigger, or if the schedules row
          hasn't caught up yet with a just-installed version.
        type: string
      schema_version:
        type: integer
      source:
        type: string
      steps:
        items:
          $ref: '#/definitions/api.workflowStepDetail'
        type: array
      trigger_cron:
        description: |-
          TriggerCron and TriggerOnMissed are only set when TriggerType is
          "schedule" — see internal/workflow.Trigger (ADR-0035).
        type: string
      trigger_on_missed:
        type: string
      trigger_type:
        type: string
      version:
        type: integer
    type: object
  api.workflowInputDetail:
    properties:
      default: {}
      description:
        type: string
      enum:
        items:
          type: string
        type: array
      name:
        type: string
      required:
        type: boolean
      type:
        type: string
    type: object
  api.workflowStepDetail:
    properties:
      binding_name:
        description: |-
          BindingName is the logical name this step's Connector expression
          refers to (the "db" in "${{ bindings.db }}"), populated only when
          Connector is exactly that shape — see bindingName.
        type: string
      connector:
        type: string
      connector_type:
        description: |-
          ConnectorType is the connector type inferred for BindingName, so a
          client can offer a <select> of matching connectors instead of a
          free-text id — see bindingConnectorType. Left empty when it can't be
          inferred (no owning plugin found, or that plugin declares more than
          one connector type).
        type: string
      id:
        type: string
      uses:
        type: string
      with:
        additionalProperties: {}
        type: object
    type: object
  api.workflowSummary:
    properties:
      id:
        type: string
      installed_at:
        type: string
      version:
        type: integer
    type: object
info:
  contact: {}
  description: |-
    Public HTTP API for the Patchcord agent (vision document, section 10.1). Covers workflow triggering and run observability, connector CRUD and testing, application hosting, and a read-only plugin catalog listing; the rest of the vision document's API surface (actions, full plugin management) is not implemented yet.
    Every route marked with a lock icon below requires "Authorization: Bearer <admin token>" — but only once at least one admin token has been created (`patchcord auth token create`); a fresh agent answers every request unauthenticated, exactly as before this existed (ADR-0036).
  title: Patchcord Agent API
  version: "1"
paths:
  /apps:
    get:
      description: Returns every installed application and the permissions its sessions
        are limited to.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.appSummary'
            type: array
      security:
      - BearerAuth: []
      summary: List installed applications
      tags:
      - apps
  /apps/{id}/sessions:
    post:
      description: 'Issues a new session for the named application, limited to its
        manifest''s declared permissions. The token is a bearer credential: pass it
        as "Authorization: Bearer <token>" to POST /workflows/{id}/run.'
      parameters:
      - description: App id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.appSessionResponse'
        "404":
          description: app not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Issue an application session
      tags:
      - apps
  /connectors:
    get:
      description: Returns every recorded connector, ordered by id. Never includes
        a resolved secret value — only each secret reference's logical type/key.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.connectorSummary'
            type: array
      security:
      - BearerAuth: []
      summary: List connectors
      tags:
      - connectors
    post:
      consumes:
      - application/json
      description: 'Creates a new connector: a persistent, named configuration for
        accessing an external system. type must match a connector type declared by
        an installed plugin''s manifest.'
      parameters:
      - description: Connector to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.createConnectorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.connectorSummary'
        "400":
          description: malformed request body, or the connector is invalid (empty
            id/type, unknown type, unsupported secret reference type)
          schema:
            type: string
        "409":
          description: a connector with this id already exists
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a connector
      tags:
      - connectors
  /connectors/{id}:
    delete:
      description: Removes a connector. There is no update endpoint (ADR-0020) — a
        caller that wants to change one recreates it, which is exactly what the dashboard's
        "Modifier" does under the hood.
      parameters:
      - description: Connector id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: deleted
        "404":
          description: connector not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a connector
      tags:
      - connectors
    get:
      description: Returns one connector's configuration and secret references (never
        a resolved secret value).
      parameters:
      - description: Connector id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.connectorSummary'
        "404":
          description: connector not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a connector
      tags:
      - connectors
  /connectors/{id}/test:
    post:
      description: Resolves the connector's configuration and secrets, then asks the
        installed plugin that declares its type to actually attempt a connection.
        Unlike GET /connectors/{id}, this calls into a live plugin process. A connection
        attempt that runs but fails (e.g. wrong password) is reported as ok=false,
        not an HTTP error.
      parameters:
      - description: Connector id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.connectorTestResponse'
        "404":
          description: connector not found
          schema:
            type: string
        "500":
          description: no connector tester configured, a secret failed to resolve,
            or the plugin call itself failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Test a connector
      tags:
      - connectors
  /plugins:
    get:
      description: Returns every installed plugin and the connector types/action ids
        its manifest declares — enough for a client to build a connector-type picker
        without a copy of the CLI's own catalog logic.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.pluginSummary'
            type: array
      security:
      - BearerAuth: []
      summary: List installed plugins
      tags:
      - plugins
  /runs:
    get:
      description: Returns every recorded run, most recently created first.
      parameters:
      - description: Restrict to runs of this workflow id
        in: query
        name: workflow_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.runSummary'
            type: array
      security:
      - BearerAuth: []
      summary: List runs
      tags:
      - runs
  /runs/{id}:
    get:
      description: Returns one run's current status, outputs and steps — the counterpart
        to POST /workflows/{id}/run's immediate response, letting a client poll a
        run or fetch its final result once GET /runs/{id}/events has closed.
      parameters:
      - description: Run id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.runSummary'
        "404":
          description: run not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a run
      tags:
      - runs
  /runs/{id}/cancel:
    post:
      description: Marks a run still queued or running as cancelled. Does not interrupt
        a run actively executing within this same process — it only flips its recorded
        status; an in-flight step keeps running until its next persistence checkpoint,
        where it then stops cleanly.
      parameters:
      - description: Run id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.runSummary'
        "404":
          description: run not found
          schema:
            type: string
        "409":
          description: run has already finished
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Cancel a run
      tags:
      - runs
  /runs/{id}/events:
    get:
      description: Streams a run's status changes, and its steps', as Server-Sent
        Events until the run reaches a terminal status or the client disconnects.
        The "event:" field is "run.<status>" or "step.<status>"; "data:" is a JSON-encoded
        event payload.
      parameters:
      - description: Run id
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: text/event-stream
          schema:
            type: string
        "404":
          description: run not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Stream a run's events
      tags:
      - runs
...
  /workflows:
    get:
      description: Returns every installed workflow version, most recently installed
        first. The same workflow id can appear more than once if multiple versions
        are installed (workflows are immutable once published — installing a new version
        never replaces an old one).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.workflowSummary'
            type: array
      security:
      - BearerAuth: []
      summary: List installed workflows
      tags:
      - workflows
  /workflows/{id}:
    get:
      description: 'Returns one workflow version''s full definition: its steps and
        raw YAML source. Defaults to the latest installed version.'
      parameters:
      - description: Workflow id
        in: path
        name: id
        required: true
        type: string
      - description: Workflow version (defaults to the latest installed one)
        in: query
        name: version
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.workflowDetail'
        "400":
          description: version is not a valid integer
          schema:
            type: string
        "404":
          description: workflow not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a workflow version's definition
      tags:
      - workflows
  /workflows/{id}/run:
    post:
      consumes:
      - application/json
      description: Starts a new run of the latest installed version of the named workflow
        and returns immediately with its id and initial status — it does not wait
        for any step to execute. Watch GET /runs/{id}/events or poll GET /runs/{id}
        for progress and the final result.
      parameters:
      - description: Workflow id
        in: path
        name: id
        required: true
        type: string
      - description: Inputs and connector bindings (both optional, default to none)
        in: body
        name: body
        schema:
          $ref: '#/definitions/api.runWorkflowRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/api.runSummary'
        "400":
          description: malformed request body, or inputs don't satisfy the workflow's
            declared schema
          schema:
            type: string
        "404":
          description: workflow not found
          schema:
            type: string
        "500":
          description: no action executor configured, or a persistence failure
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Trigger a workflow run
      tags:
      - workflows
securityDefinitions:
  BearerAuth:
    description: Admin token, if any has been created (see the top-level description).
      Pass as "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"